Flatten route groups so auth and chrome both read off the group name.
Replace the nested (site)/(protected) tree with four siblings inside the existing (app) root layout: (public), (learn), (dashboard), (player). URLs are unchanged. (player) stays a sibling of (learn) so a CMS editor can preview without a learner session.
