+++ docs/MIGRATION_GUIDE.md
@@ -83,7 +83,7 @@ const NewLessonPlayer = ({ slug, lesson }) => { /* hooks, state, JSX */ };
 
 | Old (react-router) | New (App Router) |
 |---|---|
-| `<Route path="/watch">` in `App.tsx` | A folder: `src/app/(app)/(site)/(protected)/watch/page.tsx` |
+| `<Route path="/watch">` in `App.tsx` | A folder: `src/app/(app)/(learn)/watch/page.tsx` |
 | `useNavigate()` → `navigate("/x")` | `useRouter()` from `next/navigation` → `router.push("/x")` |
 | `navigate(-1)` | `router.back()` |
 | `useParams()` | `useParams()` from `next/navigation`, or a prop from the server page |
@@ -94,20 +94,21 @@ const NewLessonPlayer = ({ slug, lesson }) => { /* hooks, state, JSX */ };
 **Route groups** are the new idea. A folder in `(parens)` doesn't appear in the URL — it exists to attach a layout or a guard:
 
 ```
-(app)/(site)/(protected)/watch/page.tsx   →   /watch
+(app)/(learn)/watch/page.tsx   →   /watch
 ```
 
-That URL is just `/watch`. `(site)` gives it the Header and Footer; `(protected)` makes it require a session. This replaces the old `<RequireAuth>` wrapper *and* the `hideHeader`/`hideFooter` pathname string-matching in `App.tsx`.
+That URL is just `/watch`. `(learn)` gives it the Header and Footer and requires a session. This replaces the old `<RequireAuth>` wrapper *and* the `hideHeader`/`hideFooter` pathname string-matching in `App.tsx`.
 
 The outermost pair splits the app from the CMS: everything the site serves lives under `(app)`, and Payload's admin and API live under `(payload)`. Inside `(app)`:
 
 | Group | Chrome | Access |
 |---|---|---|
-| `(site)` | Header + Footer | public |
-| `(site)/(public-only)` | Header + Footer | **signed-out only** — signed-in users go to `/new-lessons` |
-| `(site)/(protected)` | Header + Footer | **signed-in only** |
+| `(public)` | Header + Footer | public |
+| `(learn)` | Header + Footer | signed-in only |
 | `(dashboard)` | Header only | signed-in only |
-| `(player)` | none | signed-in only |
+| `(player)` | none | learner session, or a CMS editor previewing |
+
+`/auth` lives in `(public)` and guards itself. A layout is never given `searchParams`, so a signed-in visitor's `from` destination cannot be read in a group layout.
 
 `src/app/(payload)/` is generated by Payload — it serves `/admin` and Payload's REST and GraphQL APIs, and is not meant to be hand-edited.
 
@@ -118,7 +119,7 @@ The outermost pair splits the app from the CMS: everything the site serves lives
 | JWT in `localStorage` | httpOnly cookie — **not readable from JS** |
 | Axios interceptor attaching `Authorization` | Nothing. Same-origin fetch sends the cookie |
 | `isAuthed()` (checks a string exists) | `useSession()` in client code; `getSession()` on the server |
-| `<RequireAuth>` wrapper (UI-only) | `(protected)` route group — enforced **server-side** |
+| `<RequireAuth>` wrapper (UI-only) | `(learn)` route group — enforced **server-side** |
 | `POST /api/auth/login` | `signIn.email({ email, password })` |
 | `POST /api/auth/signup` | `signUp.email({ ... })` |
 | `GET /api/auth/me` | `useSession()` |
@@ -208,7 +209,7 @@ The Docker image bakes the model in at build time so cold starts don't re-downlo
 
 ## 5. Common tasks
 
-**Add a page.** Create `src/app/(app)/(site)/(protected)/thing/page.tsx`. Pick the group by what the page needs: chrome and access. If it needs data, fetch it in that file and pass it to a `"use client"` component in `src/pages-client/`.
+**Add a page.** Create `src/app/(app)/(learn)/thing/page.tsx` (or `(public)` / `(dashboard)` / `(player)`, matching chrome and access). If it needs data, fetch it in that file and pass it to a `"use client"` component in `src/pages-client/`.
 
 **Add an API endpoint.** `src/app/(app)/api/thing/route.ts`, exporting `GET`/`POST`. Start every authenticated handler with:
