+++ src/app/(app)/(dashboard)/dashboard/page.tsx
@@ -1,5 +1,5 @@
-import Dashboard from "../../../../pages-client/Dashboard";
-import { listLessons } from "../../../../lib/content/content";
+import { listLessons } from "@/lib/content/content";
+import Dashboard from "@/pages-client/Dashboard";
 
 export default async function Page() {
   const allLessons = await listLessons();
