+++ src/app/(app)/(dashboard)/layout.tsx
@@ -1,5 +1,5 @@
-import { requireSession } from "../../../lib/session";
-import SiteChrome from "../../../components/SiteChrome";
+import { requireSession } from "@/lib/session";
+import SiteChrome from "@/components/SiteChrome";
 
 // Header but no Footer — matches the CRA app, which hid the Footer on /dashboard.
 export default async function DashboardLayout({ children }: { children: React.ReactNode }) {
