+++ src/app/(app)/(learn)/profile/page.tsx
@@ -1,0 +1,5 @@
+import Profile from "@/pages-client/Profile";
+
+export default function Page() {
+  return <Profile />;
+}
