+++ src/app/(app)/(learn)/talk/page.tsx
@@ -1,0 +1,5 @@
+import Talk from "@/pages-client/Talk";
+
+export default function Page() {
+  return <Talk />;
+}
