+++ src/app/(app)/(learn)/watch/page.tsx
@@ -1,0 +1,5 @@
+import Watch from "@/pages-client/Watch";
+
+export default function Page() {
+  return <Watch />;
+}
