+++ src/app/(app)/(player)/lessons/[slug]/page.tsx
@@ -1,14 +1,15 @@
 import { redirect } from "next/navigation";
-import LessonRunner from "../../../../../pages-client/LessonRunner";
-import LessonPreview from "../../../../../pages-client/preview/LessonPreview";
+
 import {
   getDraftLesson,
   getDraftNextHref,
   getLessonBySlug,
   getNextLessonHref,
-} from "../../../../../lib/content/content";
-import { getShuffleIdentity } from "../../../../../lib/progress-server";
-import { getPreviewEditor } from "../../../../../lib/session";
+} from "@/lib/content/content";
+import { getShuffleIdentity } from "@/lib/progress-server";
+import { getPreviewEditor } from "@/lib/session";
+import LessonPreview from "@/pages-client/preview/LessonPreview";
+import LessonRunner from "@/pages-client/LessonRunner";
 
 /*
  * The one lesson route, either format. Was two routes (/lesson/[lessonId],
