+++ src/app/(app)/(public)/characters/[slug]/page.tsx
@@ -1,0 +1,5 @@
+import CharInfo from "@/pages-client/CharInfo";
+
+export default function Page() {
+  return <CharInfo />;
+}
