+++ src/app/(app)/(public)/forgot-password/page.tsx
@@ -1,0 +1,6 @@
+import ForgotPassword from "@/pages-client/ForgotPassword";
+
+// Public, matching the CRA route — it was never behind the PublicOnly guard.
+export default function Page() {
+  return <ForgotPassword />;
+}
