+++ src/app/(app)/(public)/funfacts/page.tsx
@@ -1,0 +1,5 @@
+import FunFacts from "@/pages-client/FunFacts";
+
+export default function Page() {
+  return <FunFacts />;
+}
