+++ src/app/(app)/(public)/gallery/page.tsx
@@ -1,0 +1,5 @@
+import Gallery from "@/pages-client/Gallery";
+
+export default function Page() {
+  return <Gallery />;
+}
