+++ src/app/(app)/(public)/layout.tsx
@@ -1,0 +1,5 @@
+import SiteChrome from "@/components/SiteChrome";
+
+export default function PublicLayout({ children }: { children: React.ReactNode }) {
+  return <SiteChrome>{children}</SiteChrome>;
+}
