+++ src/app/(app)/(public)/page.tsx
@@ -1,0 +1,5 @@
+import Home from "@/pages-client/Home";
+
+export default function Page() {
+  return <Home />;
+}
