+++ src/app/(app)/(public)/stories/page.tsx
@@ -1,0 +1,5 @@
+import Stories from "@/pages-client/Stories";
+
+export default function Page() {
+  return <Stories />;
+}
