--- src/app/(app)/(site)/(protected)/profile/page.tsx
@@ -1,5 +1,0 @@
-import Profile from "../../../../../pages-client/Profile";
-
-export default function Page() {
-  return <Profile />;
-}
