--- src/app/(app)/(site)/(protected)/talk/page.tsx
@@ -1,5 +1,0 @@
-import Talk from "../../../../../pages-client/Talk";
-
-export default function Page() {
-  return <Talk />;
-}
