--- src/app/(app)/(site)/(protected)/watch/page.tsx
@@ -1,5 +1,0 @@
-import Watch from "../../../../../pages-client/Watch";
-
-export default function Page() {
-  return <Watch />;
-}
