--- src/app/(app)/(site)/characters/[slug]/page.tsx
@@ -1,2 +1,0 @@
-import CharInfo from "../../../../../pages-client/CharInfo";
-export default function Page() { return <CharInfo />; }
