--- src/app/(app)/(site)/forgot-password/page.tsx
@@ -1,6 +1,0 @@
-import ForgotPassword from "../../../../pages-client/ForgotPassword";
-
-// Public, matching the CRA route — it was never behind the PublicOnly guard.
-export default function Page() {
-  return <ForgotPassword />;
-}
