--- src/app/(app)/(site)/funfacts/page.tsx
@@ -1,2 +1,0 @@
-import FunFacts from "../../../../pages-client/FunFacts";
-export default function Page() { return <FunFacts />; }
