--- src/app/(app)/(site)/gallery/page.tsx
@@ -1,2 +1,0 @@
-import Gallery from "../../../../pages-client/Gallery";
-export default function Page() { return <Gallery />; }
