--- src/app/(app)/(site)/layout.tsx
@@ -1,5 +1,0 @@
-import SiteChrome from "../../../components/SiteChrome";
-
-export default function SiteLayout({ children }: { children: React.ReactNode }) {
-  return <SiteChrome>{children}</SiteChrome>;
-}
