--- src/app/(app)/(site)/page.tsx
@@ -1,5 +1,0 @@
-import Home from "../../../pages-client/Home";
-
-export default function Page() {
-  return <Home />;
-}
