--- src/app/(app)/(site)/stories/page.tsx
@@ -1,2 +1,0 @@
-import Stories from "../../../../pages-client/Stories";
-export default function Page() { return <Stories />; }
